fix : Amount 타임 Long으로 수정 - Integer 타입으로 사용했을때 오버플로우 가능성 있음.

+++ src/main/java/com/example/WonkaoTalk/domain/order/entity/OrderItem.java
@@ -39,7 +39,7 @@ public class OrderItem {
   private String optionSummary;
 
   @Column(name = "product_amount", nullable = false)
-  private Integer productAmount;
+  private Long productAmount;
 
   @Column
   private Integer quantity;
@@ -49,7 +49,7 @@ public class OrderItem {
       ProductVariant productVariant,
       String productName,
       String optionSummary,
-      Integer productAmount,
+      Long productAmount,
       Integer quantity
   ) {
     this.order = order;
@@ -65,7 +65,7 @@ public class OrderItem {
       ProductVariant productVariant,
       String productName,
       String optionSummary,
-      Integer productAmount,
+      Long productAmount,
       Integer quantity
   ) {
     return new OrderItem(
